Add OpenAPI Terraform provider
The OpenAPI terraform plugin is a powerful terraform provider that enables REST APIs documented with the OpenAPI spec to be consumed via Terraform using infrastructure as code (IaC).

The provider is able to configure itself at runtime based on an OpenAPI document file containing the definitions of the APIs exposed. The dynamic nature of this provider is what makes it very flexible and convenient for service providers as subsequent upgrades to their APIs will not require new compilations of this provider. The service provider's APIs are discovered on the fly and therefore the service providers can focus on their services rather than the tooling around it (e,g: having to develop/maintain/support a custom terraform provider for the API).